+++ pkg/agent/as3/as3Manager_test.go
@@ -46,8 +46,8 @@ var _ = Describe("AS3Manager Tests", func() {
 	var mockMgr *mockAS3Manager
 	BeforeEach(func() {
 		mockMgr = newMockAS3Manager(&Params{
-			As3Version: "3.25.0",
-			As3Release: "3.25.0-3",
+			As3Version:       "3.25.0",
+			As3Release:       "3.25.0-3",
 			As3SchemaVersion: "3.25.0",
 		})
 	})